pyproject.toml
[tool.poetry]
name = "stackfordev"
version = "0.1.0"
description = "Generate tailored Dockerfiles for development environments"
authors = ["ZisisKostakakisGithubPersonal <dev@zisiskostakakis.com>"]
readme = "README.md"
packages = [{include = "src"}]
keywords = ["docker", "dockerfile", "development", "devcontainer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
]

[tool.poetry.scripts]
stackfordev = "src.cli.main:cli"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.2"
click = "^8.1"
rich = "^13.0"
httpx = "^0.27"

[tool.poetry.group.lambda.dependencies]
boto3 = "^1.35.62"
python-dotenv = "^1.0.1"
boto3-stubs = {extras = ["essential"], version = "^1.35.62"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
respx = "^0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
